Move `getHSLColorCSSVars` to a separate module
