+++ components/NotionTypes/Bookmark/link.module.css
@@ -1,10 +1,10 @@
 .link {
-    color: var(--guide-color-hsl);
+    color: var(--guides-color);
     text-decoration: underline;
     text-decoration-thickness: 1.5px;
     text-underline-offset: 4px;
-    text-decoration-color: var(--guide-color-hsl);
-    -webkit-text-decoration-color: var(--guide-color-hsl);
+    text-decoration-color: var(--guides-color);
+    -webkit-text-decoration-color: var(--guides-color);
     transform: .15s ease;
 }
