+++ consts/defaultColor.js
@@ -1,0 +1,1 @@
+export const DEFAULT_COLOR = '#f5f8fb'
