+++ pages/index.jsx
@@ -1,9 +1,6 @@
 import React from 'react'
-
 import { getTree } from '../api/apiPage'
-
 import parseManualsPreview from '../utils/manualsPreviewParser'
-
 import MainPage from '../components/MainPage/MainPage'
 
 export default function Home({ pageTitle, manualsPreview }) {
