+++ pages/manuals/[[...pageUrl]].jsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useState, createContext } from 'react'
 import { useRouter } from 'next/router'
 import { useMediaQuery } from 'react-responsive'
+import { getHSLColorCSSVars } from '../../utils/getCSSVarColorHSL'
+import { getTree, getPage } from '../../api/apiPage'
 import TableOfContents from '../../components/TableOfContents/TableOfContents'
 import ManualPage from '../../components/ManualPage/ManualPage'
-import { getTree, getPage } from '../../api/apiPage'
 import tp from '../../utils/typograf/typograf.config'
 import getManualToc from '../../utils/getManualToc'
 import { MANUAL_INDEX_PAGE } from '../../consts/manuals'
 import { PageToolbar } from '../../components/Toolbar/PageToolbar'
 import HamburgerMenu from '../../components/HamburgerMenu/HamburgerMenu'
-import { getHSL } from '../../utils/getHSL'
 
 import styles from './page.module.css'
 
@@ -90,17 +90,12 @@ function GetPage({
         anchorLinks = anchorLinks.filter((l) => l?.id >= 0)
     }
 
-    const { h, s, l } = getHSL(catalogColor)
     return (
         <>
             <div
                 style={{
-                    '--guide-color': `hsl(${h} ${s} ${l})`,
-                    '--guide-color-hsl': `${h} ${s} ${l}`,
-                    '--guide-color-h': h,
-                    '--guide-color-s': s,
-                    '--guide-color-l': l,
-                    'counter-reset': `page-chapter ${pageIndex}`,
+                    ...getHSLColorCSSVars(catalogColor),
+                    counterReset: `page-chapter ${pageIndex}`,
                 }}
             >
                 <HamburgerMenu state={isOpen} changeState={setIsOpen} />
